Adds basic potentiometer control functionality.
The potentiometer is connected to pin A0, with 5V across it. The reading comes in using analogRead(), and creates a value between 0 and 1023. These values are left to become the main delay with a maximum of a 1 second delay. The value is then divided by 5 to become the delay between "pixels" or dots that the LEDs make when flashing.